requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-filter==23.5
